+++ app/(landing)/ConceptsCard.tsx
