+++ app/(landing)/HoverCard.tsx
