+++ app/(landing)/Projects.tsx
@@ -1,5 +1,5 @@
 import {ReactNode} from 'react';
-import Section, {SectionHeading} from './Section';
+import Section, {SectionHeading} from '../../components/Section';
 import Project from './Project';
 import Gist from './Gist';
 import {GoMarkGithub} from 'react-icons/go';
