+++ app/concepts/gunn-gatm/page.tsx
@@ -1,4 +1,3 @@
-import {useEffect} from 'react';
 import Head from 'next/head';
