+++ app/not-found.tsx
@@ -1,17 +1,17 @@
-import Head from 'next/head';
-import Link from "next/link";
+import {Metadata} from 'next';
+import Link from 'next/link';
 import PageHeading from '../components/PageHeading';
 import {IoMdCloseCircle} from 'react-icons/io';
 
 
+export const metadata: Metadata = {
+    title: '404 | ky28059.github.io',
+    description: 'Page not found.'
+}
+
 export default function PageNotFound() {
     return (
         <div className="h-screen flex items-center justify-center">
-            <Head>
-                <title>404 | ky28059.github.io</title>
-                <meta name="description" content="Page not found." />
-            </Head>
-
             <main className="relative pl-14">
                 <IoMdCloseCircle className="absolute left-0 top-2 text-5xl text-grapefruit" />
                 <PageHeading className="mb-5">404.</PageHeading>
