+++ package.json
@@ -2,13 +2,11 @@
   "private": true,
   "scripts": {
     "dev": "next dev",
-    "build": "next build",
-    "start": "next start",
-    "export": "next export"
+    "build": "next build"
   },
   "dependencies": {
     "luxon": "^3.3.0",
-    "next": "^13.2.4",
+    "next": "^13.3.0",
     "next-plugin-preval": "^1.2.5",
     "react": "^18.2.0",
     "react-dom": "^18.2.0",
@@ -18,12 +16,12 @@
   },
   "devDependencies": {
     "@types/luxon": "^3.2.0",
-    "@types/node": "^18.15.5",
-    "@types/react": "^18.0.28",
+    "@types/node": "^18.15.11",
+    "@types/react": "^18.0.33",
     "@types/react-syntax-highlighter": "^15.5.2",
-    "autoprefixer": "^10.2.6",
-    "postcss": "^8.3.5",
-    "tailwindcss": "^3.0.13",
+    "autoprefixer": "^10.4.14",
+    "postcss": "^8.4.21",
+    "tailwindcss": "^3.3.1",
     "typescript": "^5.0.2"
   }
 }
