--- pages/_app.tsx
@@ -1,8 +1,0 @@
-import {AppProps} from 'next/app';
-
-import '../scss/index.scss';
-
-
-export default function App({ Component, pageProps }: AppProps) {
-    return <Component {...pageProps} />;
-}
