--- pages/concepts/codeforces/contests.tsx
@@ -1,13 +1,0 @@
-import Head from 'next/head';
-import CfLayout from '../../../components/concepts/CfLayout';
-
-
-export default function Contests() {
-    return (
-        <CfLayout>
-            <Head>
-                <title>Contests | Codeforces</title>
-            </Head>
-        </CfLayout>
-    )
-}
