--- pages/concepts/codeforces/groups.tsx
@@ -1,13 +1,0 @@
-import Head from 'next/head';
-import CfLayout from '../../../components/concepts/CfLayout';
-
-
-export default function Groups() {
-    return (
-        <CfLayout>
-            <Head>
-                <title>Groups | Codeforces</title>
-            </Head>
-        </CfLayout>
-    )
-}
